+++ src/domain/core/src/jobs/mod.rs
@@ -7,17 +7,9 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use crate::*;
-
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-#[async_trait::async_trait]
-pub trait TaskExecutor: Sync + Send {
-    /// Runs the executor main loop
-    async fn run(&self) -> Result<(), InternalError>;
-
-    /// Runs single task only, blocks until it is available (for tests only!)
-    async fn run_single_task(&self) -> Result<(), InternalError>;
-}
+pub const JOB_KAMU_CORE_DATASET_OWNERSHIP_INITIALIZER: &str =
+    "dev.kamu.domain.core.DatasetOwnershipInitializer";
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////