+++ src/utils/init-on-startup/src/lib.rs
@@ -7,16 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-#![feature(error_generic_member_access)]
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-mod entities;
-mod jobs;
-mod repos;
-mod services;
-mod testing;
+mod init_on_startup;
 
-pub use entities::*;
-pub use jobs::*;
-pub use repos::*;
-pub use services::*;
-pub use testing::*;
+pub use init_on_startup::*;